refactor(emacs): use out-of-store symlinks for config files
This change introduces a `dog.dotfilesPath` Home Manager option for
specifying an absolute path to the dotfiles repo. Out-of-store symlinks
are created for all Doom-related config and snippets, ensuring user
Emacs configuration is always sourced from, and editable in, the
git-managed dotfiles.

In practice, this removes the need for rebuilding nix files when
updating emacs config files.